refactor(console): billing UI refinements from design review
- Nav: rename Billing sub-item "Billing" → "Overview" and "Spending Limits"
  → "Limit Tiers" (match the mockup's submenu).
- Remove the Notifications menu, route, and page. Fold the billing alert
  (spend-alert threshold) into the Overview's "Estimated this cycle" card,
  next to the number it watches (get/save_billing_settings). The read-only
  feed + per-event delivery toggles are dropped with the menu.
- Merge Billing contact + Tax & compliance into one "Billing contact & tax"
  card (email · address · tax region · GSTIN, single edit pencil).
- Drop the now-unused notificationTheme helper.

vue-tsc clean; vite build green.
